fix(ios): await Matrix room metadata
